EVAKA-4161 Refactored AttendanceChildPage to just /child-info/ChildInfo.tsx
